looking for the reason things get slow over 1K path tiles
