+++ src/bundle.rs
@@ -4,9 +4,8 @@ use amethyst::{
     error::Error,
 };
 
-use crate::systems::BallSystem;
-use crate::systems::PathSystem;
-use crate::systems::VideographerSystem;
+use crate::systems::{BallSystem, PathSystem, PathTileSystem, VideographerSystem};
+
 #[derive(Default)]
 pub struct WanderballBundle;
 
@@ -18,6 +17,7 @@ impl<'a, 'b> SystemBundle<'a, 'b> for WanderballBundle {
     ) -> Result<(), Error> {
         builder.add(BallSystem, "ball_system", &["input_system"]);
         builder.add(PathSystem, "path_system", &[]);
+        builder.add(PathTileSystem, "path_tile_system", &[]);
         builder.add(VideographerSystem, "videographer_system", &[]);
         Ok(())
     }
