+++ src/systems/mod.rs
@@ -1,5 +1,6 @@
 pub use self::ball::BallSystem;
 pub use self::path::PathSystem;
+pub use self::path::PathTileSystem;
 pub use self::videographer::VideographerSystem;
 
 pub mod ball;
