+++ src/util.rs
@@ -20,11 +20,7 @@ pub fn _point_in_rect(point: &Point, rect: &Rectangle) -> bool {
     point.x >= rect.left && point.x <= rect.right && point.y >= rect.bottom && point.y <= rect.top
 }
 
-pub fn _point_on_edge_of_rect(
-    point: Point,
-    rect_center: Point,
-    dist_to_edge: f32,
-) -> Option<Side> {
+pub fn _point_on_edge_of_rect(point: Point, rect_center: Point, dist_to_edge: f32) -> Option<Side> {
     if point.x >= rect_center.x + dist_to_edge {
         Some(Side::Right)
     } else if point.x <= rect_center.x - dist_to_edge {
